fix(benchmark): record the sampling temperature as double end-to-end
An answer-variance run reported samplingTemperature 0.699999988079071 in its
DTO and its CSV: the request DTO, the frozen sampling snapshot and the freeze
service all carried a float, and copying it onto the run's double column
widened the float's nearest value rather than the number the operator typed.
Both request DTOs, BenchmarkSamplingSnapshotV1.Temperature and the freeze
service's temperature constants and helpers are double now.

No migration or entity change was needed: benchmark_runs.sampling_temperature
was already AddColumn<double> / REAL and the entity already double? — the
widening happened upstream of persistence, so AddBenchmarkRunRepeatMode is
untouched.

The sampler still narrows to float at the provider boundary (SamplingOptions,
and the judge policy's own sampling projection), which is what the wire has
always carried. The frozen payload's BYTES are unchanged — 0.7 serializes as
0.7 either way — so no stored snapshot re-hashes differently and the v1
compatibility test stays green. OpenAPI moves format float to double on
answerVarianceTemperature only; the generated TypeScript is byte-identical
because both map to number.

+++ XE-Local-AI-Engine.Client.Application/Services/Benchmarks/BenchmarkRunFreezeService.cs
@@ -45,7 +45,7 @@ public sealed record BenchmarkRunStartRequest(
     int RepeatCount = 1,
     bool Warmup = false,
     BenchmarkRepeatMode RepeatMode = BenchmarkRepeatMode.Throughput,
-    float? AnswerVarianceTemperature = null);
+    double? AnswerVarianceTemperature = null);
 
 /// <summary>
 ///     Work one launch REQUEST does once and every cell of it would otherwise repeat: the llama-server capability
@@ -190,10 +190,10 @@ public sealed class BenchmarkRunFreezeService(
     ///     default: high enough that repeats actually diverge, low enough that the divergence is still the model
     ///     answering rather than wandering.
     /// </summary>
-    public const float DefaultAnswerVarianceTemperature = 0.7f;
+    public const double DefaultAnswerVarianceTemperature = 0.7d;
 
     /// <summary>The ceiling the chat sampling UI already enforces.</summary>
-    public const float MaxAnswerVarianceTemperature = 2f;
+    public const double MaxAnswerVarianceTemperature = 2d;
 
     public async Task<IReadOnlyList<BenchmarkRunRecord>> StartAsync(BenchmarkRunStartRequest request,
         BenchmarkFreezeScope? scope = null,
@@ -393,7 +393,7 @@ public sealed class BenchmarkRunFreezeService(
     /// </summary>
     private static (int RepeatIndex, BenchmarkSamplingSnapshotV1 Sampling) SamplingFor(BenchmarkSamplingSnapshotV1 deterministic,
         BenchmarkRepeatMode mode,
-        float temperature,
+        double temperature,
         int repeatIndex)
     {
         if (mode != BenchmarkRepeatMode.AnswerVariance)
@@ -413,15 +413,15 @@ public sealed class BenchmarkRunFreezeService(
     ///     The temperature an answer-variance group samples at. Throughput mode never reads it, so a value carried on
     ///     a throughput request is ignored rather than refused — the two knobs are independent on the wire.
     /// </summary>
-    private static float ResolveAnswerVarianceTemperature(BenchmarkRunStartRequest request)
+    private static double ResolveAnswerVarianceTemperature(BenchmarkRunStartRequest request)
     {
         if (request.RepeatMode != BenchmarkRepeatMode.AnswerVariance)
         {
-            return 0f;
+            return 0d;
         }
 
         var temperature = request.AnswerVarianceTemperature ?? DefaultAnswerVarianceTemperature;
-        if (temperature is <= 0f or > MaxAnswerVarianceTemperature)
+        if (temperature is <= 0d or > MaxAnswerVarianceTemperature)
         {
             throw new BenchmarkValidationException($"The answer-variance temperature must be above 0 and at most {MaxAnswerVarianceTemperature}.");
         }
